refactor actions/action creator
